+++ doc/version-history.rst
@@ -11,6 +11,7 @@ v0.2.0
 * Make enable_cmdline_state true.
 * Check if extraTerrestrialRadiationBackwards contains None and clean it.
 * Make sure that telemetry_loop is in a while loop.
+* If bad request is received, wait an amount of time to try again.
 
 v0.1.4
 ======